Set BuildInfo for unit tests and ghci rules
Summary:
There are a handful of changes in here that should be split out when everything works appropriately.
- Exposes `ld.py` as a drop-in linker instead of g++
- Updates Haskell to use that linker (with associated buck mode changes)
- Updates `haskell.py` to do the correct buildinfo linking stuff
- Updates `BuildInfoTest` to drop the binary thing as it's slightly different from fbbuild

The core problem I have now is giving `ld.py` access to `g++` via the env vars. We need to specify this in `ghc_wrapper`, but we don't split based on `cxx_platform` to know which `g++` we can go look for and link as an env var. In this example I've hardcoded gcc5, but that clearly doesn't work for other platforms. Looks like this is yet another intersection of Haskell + platforms.
I don't know how to reconcile this.

Reviewed By: andrewjcg

Differential Revision: D5588972

fbshipit-source-id: 7abe850f40ed61fdf2b503b3df55fc9ab60eeeb6