Add support for passing through platform regex and platform linker flags
Summary:
Allow passing os-specific dependencies and linker flags for a rule. Since we want to have a few restrictions still, we only allow coarse grained 'os' selections. Any deps or flags that do not match the current platform are simply omitted, rather than going through buck's platform logic. This means we don't have to parse extra deps that we aren't going to use, and keeps the query graph a bit simpler.

Platform deps were added so that we can drop some rules on osx (like the folly stacktracer that requires libunwind and elf). Linker flags were added so that we can add things like framework linkage on osx for watchman

Reviewed By: andrewjcg

Differential Revision: D5433187

fbshipit-source-id: b77d94db379051bc805a0a735e0a103b3e31555d