Support GHC rts library linking based on `hs_profile`, `hs_eventlog`, and `hs_debug` flags
Summary:
So here's one way that works 100% outside of messing around with buck. Since `rts` is special and always an "external" library, we can special-case it when generating the lib files to pass to buck. This perpetuates the hack that fbbuild used to accomplish the same thing.

Questions:
- I don't know if this messes with caching. What do you think andrewjcg?
- I force everything to use a form of the threaded runtime. Think that's a problem simonmar?

Other notes:
- This works for c++ and hs targets
- You no longer specify profiling/eventlog/debug via parameters in TARGETS files. Instead you mark them on command line with `-c fbcode.hs_<type>=true`

Reviewed By: simonmar, niteria

Differential Revision: D5721283

fbshipit-source-id: 5a6fbb69e6206dc85afd75ca202725e0e54e43f0