enable shift-exponent UBSAN check in ASAN
Summary: > -fsanitize=shift: Shift operators where the amount shifted is greater or equal to the promoted bit-width of the left-hand side or less than zero, or where the left-hand side is negative. For a signed left shift, also checks for signed overflow in C, and for unsigned overflow in C++. You can use -fsanitize=shift-base or -fsanitize=shift-exponent to check only left-hand side or right-hand side of the shift operation, respectively.

Reviewed By: meyering

Differential Revision: D5518510

fbshipit-source-id: b6de043771e59ec8140bade66a697ee436b3f396