Allow forcing the optimisation level per target
Summary:
I wanted to set -O0 for sigma/repo to make it rebuild faster, but Buck
wouldn't let me.  Not sure if this is legit or not, but the fact that
you can do it manually by adding `{-# OPTIONS_GHC -O0 #-}` to every
source file suggests that we're not making anything worse by doing
this.

Reviewed By: JonCoens

Differential Revision: D5890878

fbshipit-source-id: 1ebc78bc17f505f3769ebb1e63bb6f301a18db07