Adjust deps for gtest tests
Summary:
For gtest tests, always depend on gtest and gmock. And depend on our default runner only when not disabled.

Now that gtest and gmock ship together, and given that so much depends on gmock as well as gtest, might as well add it by default as a dep to all gtest-style tests.

Reviewed By: igorsugak

Differential Revision: D4889903

fbshipit-source-id: ed94d5256cd4e2f2df4e0dc059aa0e5fbdca3c32