always ADD_THRIFT_EXCEPTION in the thrift_library macro
Summary:
There's one obvious case where ADD_THRIFT_EXCEPTION can't be safely applied. As
we work toward creating consistency across thrift_library swift rules, the
simplest way to whitelist the one usage is to rename the option to something
that makes it clear that it's not the recommended way to do things.

Differential Revision: D5983225

fbshipit-source-id: c2e9777a79f213364b2079253d8cd2f8165d2923