if java_swift_maven_coords is present, maven_publisher_enabled=True
Summary:
This diff effectively makes the change in D6093697 have transitive effect. The
changes to `remapping/TARGETS` and `thrift_library.py` are manual. The rest
of the changes are automated, using buildozer.

The `thrift_library` rules modified in this diff are the transitive dependencies of
existing `thrift_library` rules that have `java_swift_maven_coords` specified. When
setting `maven_publisher_enabled=True`, the code added in D6007012 asserts
that all dependencies on the classpath can be mapped to `maven_coords`.

Differential Revision: D6094102

fbshipit-source-id: a2ed3a2bc34e1e3fda7c9967c6522b7fb695ff95