buckify_tp2: create missing project stubs
Summary:
This diffs makes buckify_tp2 create a missing stubs TARGETS for projects for
platforms it isn't available for.  This TARGETS contains rules which, when
dependend on and built, fail the build with a missing project error message.
With these stubs, we can now omit projects for certain platforms and build
correctly with Buck.

Reviewed By: pixelb

Differential Revision: D6082086

fbshipit-source-id: e850567ae301afb8b7cb2a6381f9143a69689785