buck macros: support generate source references to third-party rules
Summary:
Currently, when referencing a rule from third-party in `srcs` using a target
reference, it gets converted to a Buck target using the default platform.
This diff parses source path parameters and installs ones that reference tp2
rules via the `platform_srcs` parameter, and updates Python and C/C++ converters
to use this.

Reviewed By: igorsugak

Differential Revision: D6134174

fbshipit-source-id: f75dee05fd33b9f20bb7932a01dc62b34b4795bd