buck macros: make config.py usable via global import
Summary:
This diff refactors the macro library import setup a bit so that `config.py`
can be loaded via `include_defs()` and it's configuration settings read
gloabally.  This allows subsequent diffs to incrementally move parts of the
macro lib which rely on configuration into standalone macro def modules.

The diff adds a fair amount of unappealing compatiblity code to allow this
migration to be done incrementally.  It should be trivial to clean this up
once porting is complete.

Reviewed By: philipjameson

Differential Revision: D6194686

fbshipit-source-id: af8cff978f5e8ed84a52304cf8c0c841de3ca67c