thrift_library finish renaming java{,deprecated}
Summary:
For a while, the recommended approach (e.g. from folks on the thrift team, from
folks on the javafoundations team) has been to use `languages = ["java-swift"]`.
This renaming brings the code in line with the recommendation. There are no
actual changes to any builds, this is just a rename.

This is the third diff in a small stack. It removes the shims added by D6204449
and renames `thrift_java_options` to `thrift_javadeprecated_options` to
complete the rename.

Reviewed By: k21

Differential Revision: D6206118

fbshipit-source-id: 8d8a89af09d31fdb8a373a47b7f8c5d77f080dc9