rust: thrift: introduce Framing trait and use it
Summary:
This change introduces a Framing trait, which is responsible for managing
memory for memory buffers for serialized frames: DecBuf is for incoming buffers
to be decoded, and EncBuf is for encoding frames. This is intended to allow
specific framing implementations to use their own buffer types, such as
bytes::Bytes or C++-level memory buffer.

This is coupled with TProtocolFactory, which constructs instances
of protocol-specific encoders/decoders using the corresponding Framing
buffer types.

This combination subsumes tokio-io and tokio-proto, so clients clients just
need an implementation of tokio_service::Service. As a result the breakdown of
responsibility is:

Framing - in memory buffers

TProtocolFactory - encoders of requests/responses to/from a Framing buffer in a
  given protocol (Binary or Compact right now)

tokio_service::Service - either a client or server implementation. The client
  implementation actually gets a request frame sent and a corresponding reply
  returned. For servers, implemented in terms of a ServiceProcessor dispatching
  to the service implementation trait.

Reviewed By: farnz

Differential Revision: D6169708

fbshipit-source-id: bb12fdcab2541c57d1e3cb1c68a0ac8f40fd5d88