buck macros: replace clowny `include_defs` with flake8-friendly `import_macro_lib`
Summary:
This is a follow-up to D6196019. My suggestion is that we paste the 8 lines of `import_macro_lib` at the top of every file needing to do imports.

That way, we get to operate almost like a normal Python program from the point of view of scope management, and continue to benefit from flake8.

Reviewed By: philipjameson

Differential Revision: D6216779

fbshipit-source-id: 570d1ef308badea07d8abad7b6715b65b7efdbcc