`image_feature` Buck target + import to compiler Items
Summary:
This adds `image_feature` to Buck, which is a composable gadget that collects future image Items, and writes them to a JSON file.

There's also a bit of code for the compiler to import that resulting JSON.

The test just exercises that the Buck parsing + compiler import together work as expected.

Reviewed By: marcinpe

Differential Revision: D6181433

fbshipit-source-id: edbfbad89c2d3b4c2d651b953893ac9ef1c1eb7c