use the portable platform.machine() not platform.processor()
Summary:
platform.machine() is portable and populated by the kernel.
On some systems (like centos) platform.processor() is simulated
by copying the value from platform.machine(), though this doen't
happen on systems where coreutils' uname is not so adjusted.
For example on debian or if running upstream coreutils.

This patch was automatically generated with:

  sed -i 's/platform\.processor/platform.machine/g' \
    $(fbgs -ls platform.processor)

I also verified platform.machine() returns appropriately on 'aarch64'

Reviewed By: meyering

Differential Revision: D6410633

fbshipit-source-id: 38694cb0c4135d591289454e02b54f8be6993e1f