use binary-parsers instead of attoparsec
Summary:
attoparsec is overkill for thrift deserialization because it can do backtracking. The thrift protocols are very simple and don't need any backtracking, so we're better off using a less powerful, faster parsing framework.

By changing to this new underlying parsing library, we get a ~33% parsing speedup on a particularly large configerator config. Unfortunately, we increased allocations almost 2x. The reason for this is because I switched from the very fast aeson string parser to our slow custom string parser. I'm going to fix our string parser to be fast, but I want to do it in another diff.

Reviewed By: watashi

Differential Revision: D6303830

fbshipit-source-id: 1ff105f764a2393ff30307e7c730a05e8b9b37ab