Send binary_link_deps as preload dependencies for other haskell target types
Summary:
Since the `iserv` process needs allocators and whatnot to be preloaded, we need to send this dependency down for each target. We don't know at the macros-level whether we're running with `ghci` requested (which is unfortunate), so we send it down each and every time.
This will depend on (D6263634).

Reviewed By: andrewjcg

Differential Revision: D6263461

fbshipit-source-id: 638e95e043f9a6466e3372ac5f1b17f1f4c1c16d