buck macros: build with undefined symbols
Summary:
By default, the linker will not error out when linking a shared library if
there are references to undefined symbols in the constituent object files.
In our normal build flow, we instead would see these errors when linking
a top-level binary.  However, the use of shared library interfaces, which
scrub undefined symbols from the minimal shared lib deps used for
linking, means that we now only see these undefined reference errors at
runtime.

This diff adds the `-Wl,--no-undefined` flag to the linker flags when linking
shared libraries, so that we instead see errors when building the shared
libs.

Reviewed By: pixelb

Differential Revision: D6426369

fbshipit-source-id: 1b7070a4d67dbb1c0eeae413be803ceb700c25cf