buck macros: add a `cpp_jvm_library()` rule
Summary:
This adds a rule which is meant to wrap the fbcode platform JVM for C/C++ rules
to depend.  It is meant to replace the numerous amount of hard-coded references
to the system JDK found in TARGETS (D6561345, P58785788).

It should be noted that this rule doesn't actually make Buck track/hash the
underlying includes/libraries, and is just meant as a incremental step to unify
the C/C++ -> JVM flow in fbcode.

Reviewed By: k21

Differential Revision: D6561463

fbshipit-source-id: 1eca3090d6d5ea618e5e634ed403f78be10d744c