unit tests for per-repo volumes
Summary:
Our btrfs volume management will get executed as part of `buck build`. So, building an image constitutes an integration test.

However, it is really useful to be able to test the individual steps automatically, and with more care.

This here is a first step in establishing this pattern. Due to the divergence between our unit-testing / continuous integration setup, and the Python environment provided by Buck, there are a few parts that cannot effectively be unit-tested.

However, this is still a net win. For example, this unit test found the soft-lockup kernel bug mentioned in the comments.

Drive-by fix: I discovered `# pragma: no cover`, and was able to prove that we have 100% coverage the currently defined Python libraries.

Differential Revision: D6606065

fbshipit-source-id: da2fcd0761da6ef4f840d268c7770d8980b67219