Add cell for new macro lib location
Summary: Adds a cell called 'fbcode_macros' to make it easier to include macros in build files and other macro files. It also will make it easier to include them for OSS purposes. This new directory is where we'll store skylarkified macros. There was a problem with the configs before that caused havoc in other repos, so let's try this once more.

Reviewed By: mzlee

Differential Revision: D6654201

fbshipit-source-id: 5d029693c2e08bd14e3739481808ca4449074f67