subvolume uuid for built image layers
Summary:
This is nothing but a random-number generator. The resulting string is appended to the name of the subvolume directory that we use to build any given image layer.

One could argue that buck-built image layers should be versioned using the Buck rule key instead. That would be meaningful if our image compiler supported iterative rebuilds, but since it does not, it is actually safer to always give it a clean output directory.

Differential Revision: D6670812

fbshipit-source-id: afe15c371d8a237ceedc452db42ed03ba0a7a858