Buck output-able disk representation for subvolumes
Summary:
After building a btrfs subvolume, we cannot just hand it to Buck, since it cannot be treated as a regular directory, and (unless `buck-out/` is in a btrfs volume, which is not guaranteed) it cannot even be placed in `buck-out`.

So instead, we will build a JSON object that points to the actual subvolume. We will get the first variant of this object from our image builder (`from_build_buck_plumbing`). We then do a bit of processing to avoid storing anything but repo-relative paths in the JSON, and output it for real (`to_json_file`).

This is a stand-alone module for testability reasons, and also because both the compiler & the subvolume garbage-collector need it.

Reviewed By: marcinpe

Differential Revision: D6681952

fbshipit-source-id: 01b7bc0c52f8d880b743d82506b3175e9dc76981