pick a directory for image build artifacts in the source repo
Summary:
Image building produces btrfs subvolumes, rather than vanilla files. We can't expect Buck to know how to cope with those, so instead we pass it small `.json` files that point at the actual subvolumes containing the data.

The current diff picks a location for the artifacts directory, which will store the actual subvolumes, as well as some refcounting data, and potentially other image layer metadata that cannot be represented as a Buck output.

We are putting the artifacts directory in the repo root as a sibling of `buck-out` for two reasons:
 - Analogy — it helps users understand that this is similar to buck-out in purpose.
 - Resistance to rename refactors — the main alternative to the repo root would be to put the artifacts directory under `container_image/` in the source tree. However, this would result in user's repos leaking the cached subvolumes if any refactor moves `container_image/` elsewhere in the repo. Such a move would also cause all `.json` references to those subvolumes from `buck-out` to become invalid, necessitating a `buck clean`. In all, cleaning up after a move would be a manual fix that every repo user has to do separately, and so we do our best to avoid it.

Reviewed By: marcinpe

Differential Revision: D6671141

fbshipit-source-id: b4d05b4036a9b45e52d859a86ac23528a9b69f67