Add config.bzl, add simple testing framework
Summary: This adds a single entry from config.py to config.bzl. It's actually mostly adding a small framework so that we can do type-safe unittests of bzl files. This will be expanded a bit later to make sure that we get expected final-forms of rules (e.g. buck audit rules)

Reviewed By: andrewjcg

Differential Revision: D6762786

fbshipit-source-id: 54f5c3974fbc6868078834fa93ab4b95649658d8