Add metaclass that forces testing both skylark and python
Summary:
Make our test functions run both with python, and skylark with the same data. This should help us make sure that both interpreters give us the same output.

This also imports six because metaclasses changed between python 2 and 3 and I'd like it to work on both interpreters, and stay isolated from the rest of fb infra for easy OSS

Reviewed By: ttsugriy

Differential Revision: D6826805

fbshipit-source-id: caf41c156b8356187b0edda2fff3acab28fd5d1d