Adding esan-cache.
Summary:
esan-cache (efficiency sanitizer cache fragmentation) is a tool the determine hotness of fields inside structs/classes.

It instruments the compiled program to collect access counts.

If other tools under esan are added, they should be added as a different build mode.

esan is made to be an umbrella of performance monitor tools, cache-fragmentation is one of those
tools that is useful to us.

http://lists.llvm.org/pipermail/llvm-dev/2016-April/098355.html

Reviewed By: igorsugak

Differential Revision: D6844416

fbshipit-source-id: f36bccde6b71f11ff97fb1d16bf94c019d44ca00