python_* rules runtime_deps and cpp_deps
Summary:
We are trying to make deps only be code codes, so we can generate type checking
configs based on the contents of deps. Which means they need to be python_libraies.
Or known ways to produce python compiled extensions (cython, swig, cpp_python_)

So to provide for ctypes we have cpp_deps
To provide for the case where somebody just wants a target built but its not
a code dep and not packaged we have runtime_deps.

Reviewed By: ambv

Differential Revision: D6922960

fbshipit-source-id: d6666da6de8388ddf3ffcc5d8b93af039adeb501