Prevent Java thrift class configs
Summary:
Introduce a new script that forces thrift files that are supposed
to be compiled with the Apache Thrift compiler to contain a special
tag and files that are supposed to be compiled with fbthrift to not
contain this tag. This way we can guarantee that no file will be
processed by both Apache and fb thrift which would lead to
conflicting classes.

Reviewed By: ryandm

Differential Revision: D6908145

fbshipit-source-id: b1c065c1cbe9fdcded8ccf995a5886bf70d964d9