Only emit error rules if a wheel is missing on both Python versions
Summary:
There's a number of packages that only make sense on Python 2 or on Python 3.  In the former case, they are most often backports of packages that are built into Python 3.  In the latter case, it's new functionality that uses syntax and features unavailable under Python 2.  Those would fail to import anyway but since applications are using `try: import X; except ImportError: do_something_else()`, it doesn't matter that a wheel is missing.

We don't want PAR builds to fail if a PAR was available for the other Python version on the given fbcode platform.  We just acknowledge it's there and let it go.  In all likelihood everything will work just fine.  And if it won't work, the import error will be obvious to the user as to what is missing.  Checking //python/wheel/X/TARGETS will reveal a missing wheel.

- People blocked on this so landing (@[530005397:cooper])

Reviewed By: andrewjcg

Differential Revision: D6928330

fbshipit-source-id: c5afa2f5182908aac832567242f0134e9b252042