buck macros: don't use global default plaform in auxiliary version handling
Summary:
The auxiliary version support in Buck for fbcode is a hack to work around issues
with neteng and Django.  Previuosly, this would process deps at parse-time using
the configuration from the default platform, which could cause issues when trying
to build for other platforms which were configured differently.

This diff moves the processing to formatting time and only to the python rule
handlers.

Reviewed By: pixelb

Differential Revision: D6933314

fbshipit-source-id: 481bb8f878c6144af54cb23a2917b984a272270f