Switch compiler to relative imports
Summary:
It's really useful to be able to run this code both under Buck, and using the system Python.

The current setup had a split brain: the compiler considered itself the project root, but the outer utilities imported compiler code as `compiler.MODULE_NAME`. This works in Buck due to its `base_module` support, but it breaks system Python.

I could get absolute imports to work by deciding on an absolute root for the entire project, such as the `container_image` directory, but overall that seems like more hassle to evolve than relative imports, and more verbose.

Differential Revision: D6973819

fbshipit-source-id: 5fd106e9634141e1ef249befc8cf2662796fdecd