Don't pass lld '--thread-count' option
Summary:
Although lld is meant to be a drop-in replacement for the
GNU gold linker, and so should at least be able to handle
all the options that gold does, it does not (yet?) handle
the `--thread-count` option.

Add a crude check to prevent lld from being passed this
option. As the FIXME notes, the check will not work if
Clang is passed an absolute path, such as
`-fuse-ld=/path/to/lld`. Still, for now, this is sufficient
for the invocations I plan to use this with.

Reviewed By: igorsugak

Differential Revision: D7051612

fbshipit-source-id: 9a243a6f619f64d90fe9309b3d98b337fb12e28f