btrfs diff validation: parse `btrfs receive --dump`
Summary:
This is logically independent of D7105856 and D7106034, but they will connect in the following way:
 - This diff parses individual lines of `btrfs receive --dump`, but does not really understand what filesystem effects these lines have. It's just a dumb parser.
 - The other diffs are the trickiest part of a full mock of VFS.
 - Once I've put up the final diff (file & directory metadata), I will be able to have an integration test that generates a btrfs diff, applies its operations to our mock filesystem, and validates its internal state.

As discussed in the introduction to D7105856, the approach of comparing mock filesystems has a number of usability & reliability advantages over comparing live filesystems. It also serves as a layer of test coverage for ensuring that the real FS behaves as we expect.

Reviewed By: zeroxoneb, marcinpe

Differential Revision: D7118192

fbshipit-source-id: c98e4fbbf47400504c513456e10e0c5c80004a56