Decouple `IncompleteInode`s from the `Extent` -> `Chunk` transformation
Summary:
My initial naming, and reliance on inode-like objects, was ill-advised. IncompleteInodes aggregate a variety of state as we consume the btrfs send-stream. It's not appropriate to tightly couple that process to the very specific concern of producing clone-aware chunk lists.

Put another way, a subsequent diff would be far messier if I had continued operating on `IncompleteInode`s in these functions.

Reviewed By: marcinpe

Differential Revision: D7189112

fbshipit-source-id: f03c248bbeb16dcc4247e020c7df9114fce9d2d7