A more complete IncompleteInode
Summary:
The new variant of the class knows how to apply `DumpItems`, and understands all VFS metadata except for `chattr`, which is as-of-now unsupported by `btrfs send` anyway.

I purposely left out `clone` support, because that requires a lookup from `path` to the corresponding `IncompleteFile`, and that is something that is handled by `Subvolume` (next diff).

`Subvolume`, of course, depends on `IncompleteInode`, so I couldn't land them first. And in wanting to keep my changes small and well-factored, I wouldn't want the two submodules to be in one diff.

So that's why, `clone` support will land last.

Reviewed By: marcinpe

Differential Revision: D7201597

fbshipit-source-id: 5cf72631785ba829f80d39fdcc06b72dadbdecc5