always generate sanitizer configuration library per binary rule
Summary: In some cases, sandcastle will select all of the build rules in a directory in the default mode (dev) but will use them in another mode. It worked well as we always generate the same dependencies in different build moded. Since D7222284 we generate special sanitizer configuration build rule only if any of the sanitizers are enabled. That broke some of the sandcastle commands. It's unclear how many broken sandcastle commands there now. For now just enable the sanitizer configuration build rule in all modes. The impact should be minimal: the rule has no dependencies and if the sanitizer is not enable it'll be empty.

Reviewed By: meyering

Differential Revision: D7304338

fbshipit-source-id: feba8a0a70f2862f8707b48d86376a0523900df1