Make explicit that send-stream parsing must track the current volume
Summary:
`--dump` injects the current subvolume's name into some, but not all of the paths (see e.g. `link`). The actual binary send-stream only stores subvolume-relative paths.

Here, I make `parse_dump` operate as if it needs to track the current subvolume. This means that dump items can no longer be parsed in one-by-one. However, building actual filesystems also requires us to operate on a whole sequence of items, so this restriction is not a problem.

As a reward, the outputs of `parse_dump` and `parse_send_stream` can now be almost identical, and we can lose the weird `SubvolPath` abstraction I introduced earlier.

Reviewed By: marcinpe

Differential Revision: D7333519

fbshipit-source-id: 644cf0c0f1bfdf311d543b5cc01a865e9444470a