Coroutine utility `while_not_exited`
Summary:
I realized I want to be using more of the pattern of "a coroutine has business logic, and a test drives it to inspect or influence intermediate states" (the thing I'm doing in `InodeIDMap`).

The pattern is useful, but Python coroutines (aka generators + `.send` + `.throw` + `.close`) are far from intuitive.

So I rolled this tiny abstraction, and a small test suite to use as a reference so that I'm less likely to use this pattern badly.

Reviewed By: marcinpe

Differential Revision: D7390098

fbshipit-source-id: 40d8cdc471ce4c20285a7881b37b6bd1834f2bb5