Test ensuring that InodeIDMap is `deepcopy`able
Summary:
This is the sanest way of snapshotting mutable data structures in Python, assuming they don't store anything sketchy.

Writing a manual copy constructor is more code and more error-prone (I did, in the first pass at this change).

The test is important for ensuring that we continue not to store anything sketchy in the map. It's not bullet-proof, but it's definitely better than nothing.

Reviewed By: marcinpe

Differential Revision: D7369158

fbshipit-source-id: 8d8b9f22e3fe3b27b32c948d66b3acdfe567193c