Explain why `IncompleteInode` is correctly `deepcopy`able
Summary:
I need `deepcopy` for snapshotting `Subvolume`s.

Unfortunately, I do not see a productive way of writing a test of `deepcopy`ability of these classes, because the main point of concern is that copying can introduce subtle object aliasing issues.

This is why I ought to finish `deepfrozen` -- having **everything** be recursively immutable eliminates the need for `deepcopy` shenanigans, and so much more.

Reviewed By: marcinpe

Differential Revision: D7391715

fbshipit-source-id: a1591512992f5d66cde7db83f7b959d141fd1c43