Subvolume
Summary:
The brtfs emulation is built up in a few layers. So far, I've landed:
 - `Extent`
 - `IncompleteInode` subclasses, including `IncompleteFile`, which contains `Extent`s

Now I'm adding `Subvolume`, which is a path-indexed collection of `IncompleteInode`s.

Everything below `Subvolume` must be correctly `deepcopy`able to support snapshotting.

The last step will be a `SubvolumeSet` which is a UUID-indexed collection of `Subvolume`s.

Reviewed By: marcinpe

Differential Revision: D7419308

fbshipit-source-id: 862a7a846662ebd120d086b349b08c60e597856b