Add `freeze` support to `IncompleteInode`, move its `repr` logic into `Inode`
Summary:
The overall thrust of these next few diffs is to make it so that I can take a filesystem (e.g. `SubvolumeSet`) that is under construction, and `freeze` it to produce something that is recursively immutable and has all the cloned extents annotated.

The reason that I **need** things to be recursively immutable is because I did not write the clone-detection logic to support live updates. And so allowing a mutable filesystem to have annotated clones would just be a path towards much sadness.

What could be better here?
 - If I had `deepfrozen` finished up, some of the boilerplate would disappear.
 - If somebody has a clever idea for a live-updating clone-aware extent implementation, it should be possible to refactor the whole `freeze`ing thing away. Although having a read-only view seems like a helpful thing for sanity anyway.

Reviewed By: marcinpe

Differential Revision: D7452828

fbshipit-source-id: 5a95370c2689675e6468533ff47843c4367e80b0