Fix check_types for py_version = [], and fix typing_options for non libraries
Summary:
When using py_version = [2, 3] it should be possible to say check_types = True,
without raising an exception.

py_version is a list, the magic genrule target for doing `buck test`
now inherits the 'tests' options from the versioned unittests.  This means it will
contain the -typecheck target if check_types is set to true.

typing_options were not being generated for python_binary and python_unittest, unless
typing=True was set, but that option was not suppose to be valid for python_binary and python_unittest.
So now if check_types is true and its not a library then typing information will be emitted.

Reviewed By: ambv

Differential Revision: D7495741

fbshipit-source-id: 03132c26e8200904cc7d7f49dcf3be62ed1e3ff4