Non-broken implementation of InodeIDMap: rename directories, no directory hardlinks
Summary:
My prior implementation of `InodeIDMap` was completely broken for the purposes of renaming directories. Since every item stored full paths, the renaming `a/` to `b/` would result in abominations like `a/d` is a child of `b/`.

To put an end to this nonsense in the prior implementation would require rebuilding an entire subtree of the map every time we renamed a directory. Worse, since the prior one did not forbid directory hardlinks, you could imagine siblings `a/` and `b/`, which are hardlinks, with both of them linked into the same directory, etc, etc. Basically, the whole data model was kind of busted and incoherent.

To make it right, I unfortunately had to build out the linked data structure that corresponds to an actual filesystem, so this diff is a bit big. Not sure if I could've made this dramatically shorter.

Reviewed By: zeroxoneb

Differential Revision: D7513564

fbshipit-source-id: e206b0493333757ff3edbe33898fd59c98a762ee