Start porting and simplifying path logic from convert/base.py
Summary: This adds the logic for converting external_deps to actual targets. It parallels the transformation logic we had previously in base.py, but does not have the intermediate objects. If we decide we really want those objects (I'm less convinced that we do), we can always put them back, but the logic seems easier to follow here if we just go tuple -> target string

Reviewed By: ttsugriy

Differential Revision: D7785435

fbshipit-source-id: e240f68a18e462be5a233b4ab445b4e3efca6c7b