Don't extend for empty params
Summary:
When using 'extend_build_mode', only combine attributes if
necessary. For example, when extending a parent build mode
that only sets 'asan_options' to set 'cxx_flags', do not
attempt to combine 'asan_options' with the default parameter
of an empty tuple.

This fixes an issue when extending parent build modes that
set dicts with child build modes that do not set those
attributes at all. Previously an error would occur when
attempting to combine a dict with a tuple.

Reviewed By: philipjameson

Differential Revision: D8070854

fbshipit-source-id: 7d2ce4a2f25973e8e041a7f57ede11eb6ec6ee96