buck macros: add virtual `third-party-tools` cell for tp2 tools platform
Summary:
A relatively common flow for rules is to access binaries/tools from the tp2
tools platform.  As we don't really provide an effective way to do this, it's
common for really hacky solutions to pop in TARGETS in their place (e.g.
https://fburl.com/kwpk3039).

This diff adds a virtual cell, `third-party-tools`, which can be used to
in targets/location-macros to address rules wrapping tp2 binaries in the tools
platforms.

Reviewed By: pixelb

Differential Revision: D8273995

fbshipit-source-id: ccb15f7286094bf072fb070ad7cd655bf896bf1c