Keep using include_dir temporarily
Summary:
Our strategy for removing auto detection/generation from include_dir right now is:

- Add parameters to macro lib
- Update build files to set include_dir to 'None' if using versioned_header_dir or an array if no versioning is required.
- Remove old include logic from macro lib

In between 2 and 3, we need to make sure our build files don't change behavior, so imitate the old logic, by basing it off of the old logic. This will be deleted when the old include logic is removed from the macro lib.

Reviewed By: ttsugriy

Differential Revision: D8083723

fbshipit-source-id: 33d9eac76ccc9d5d3d2678113e3e7a0f21398a53