buck macros: don't use `fbcode.platform` config for D and legacy Go platforms
Summary:
This vacates use of the `fbcode.platform` config when configuring D and legacy
Go platforms and instead just uses a hardcoded default platform settings.  This
trades off a little flexibility for these cases (which I doubt will be noticeable)
to pave the way for no longer requiring `fbcode.platform_files=false` to change
the fbcode platform and to allow configuring clang opt builds on a per-project
basis.

Reviewed By: philipjameson

Differential Revision: D8310026

fbshipit-source-id: 4163c824177f51e171a1f65a93d9c563e70f6554