add nodefaultlibs option to ocaml and flowmin build
Summary:
See https://fb.facebook.com/groups/fbcode.foundation/permalink/1300384513397351/. tl;dr: The default libs are unfortunately killing us -- we spin up a lot of small processes to minify each file and they all contend on the `sysfs` spinlocks in `folly::CacheLocality` static initialization. I debugged this really carefully with Linux perf tools and know that it is a huge issue for clean www builds.

The binary I produced after this change works and the CPU is almost 100% maxed out in userspace, computing the syntax transform, whereas before the entire box was like 60-70% kernel space, so the change is working as intended.

Reviewed By: andrewjcg

Differential Revision: D8529375

fbshipit-source-id: 3d202e95315928e717bf1fa13bea9c410ac1d64f