buck macros: switch C/C++ rules to use `lang_platform_compiler_flags` (take 2)
Summary:
Switch sending compilers flags to Buck to use the `lang_platform_compiler_flags`
parameter and update compiler-specific flags to pass based on compiler-specific
platform.

This a second attempt at this diff.  The first had to be reverted as the version
bump it depended on caused issues in fbcode (which were unrelated to this change).

Reviewed By: pixelb

Differential Revision: D8554865

fbshipit-source-id: 564f5f64a0de33ca2f437a98e169655193fba327