buck macros: add plumbing to support multi-compiler family builds
Summary:
This splits the previous required global compiler setting into an optional global
setting and a default fallback setting.  If the global compiler is set (e.g. in
`dev` mode), then we maintain the original behavior where eveythhin in that build
mode will build with that compiler.  If the global compiler is not set, then
individual rules can choose between gcc and clang (support for this coming next),
otherwise falling back to the default setting.

Some features (e.g. sanitizers, coverage) require clang, and so these will continue
to a globally set compiler, and asserts in the code to check for this have been
updated to support the new structure.

Reviewed By: pixelb

Differential Revision: D8394371

fbshipit-source-id: d27973507c50cc8a480c67a36f0796332ede3961