disable null UBSAN check
Summary:
Considering the instrumentation overhead of this check disabling it for now. For some binaries the added overhead increases the binary process foot print by 2x, and causes program text interleave with the ASAN shadow memory. There are a few techniques being evaluated to move the shadow memory further into the address space. But regardless of the approach it'll cost extra CPU overhead.
The risk of the missing out on the critical null pointer use bugs is reduced by ASAN. The compiler often optimizes away, inlines or reorders the instructions making many of this violations undetectable by ASAN and unlikely to crash the program.

Reviewed By: taewookoh

Differential Revision: D8568980

fbshipit-source-id: d4862dba35674dea6d63a3d5c749ff9fcf0f5efe